GUI: Allow automatic lay-outing of ImGui windows
This is done initially when no ImGui settings are present in the config
and can also be triggered by a button or other means.

A layout builder is also added to make it easy to sketch layouts in
code. The only real limitation is that the order of splitting is fixed.
This can be tracked in the builder but is not implemented yet.